+++ homework 3/SimpleFTP/SimpleFTPClient/ClientSource/Program.cs
@@ -1,7 +1,7 @@
 ﻿using System;
 using System.Threading.Tasks;
 
-namespace Source
+namespace ClientSource
 {
     class Program
     {
